move shared schemas to own folder
