--- plugins/optimizer/coordinator/shared_schemas.py
@@ -1,125 +1,0 @@
-# Copyright 2023 QHAna plugin runner contributors.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from dataclasses import dataclass
-from typing import Optional
-
-import marshmallow as ma
-import numpy as np
-from plugins.optimizer.shared.schemas import NumpyArray
-
-from qhana_plugin_runner.api.util import MaBaseSchema
-
-
-@dataclass
-class LossResponseData:
-    loss: float
-
-
-class LossResponseSchema(MaBaseSchema):
-    loss = ma.fields.Float(required=True, allow_none=False)
-
-    @ma.post_load
-    def make_object(self, data, **kwargs):
-        return LossResponseData(**data)
-
-
-@dataclass
-class TaskStatusChanged:
-    url: str
-    status: str
-
-
-class TaskStatusChangedSchema(MaBaseSchema):
-    url = ma.fields.Url(required=True, allow_none=True)
-    status = ma.fields.String(required=True, allow_none=False)
-
-    @ma.post_load
-    def make_object(self, data, **kwargs):
-        return TaskStatusChanged(**data)
-
-
-@dataclass
-class CalcLossInputData:
-    x: np.ndarray
-    y: np.ndarray
-    x0: np.ndarray
-
-
-class CalcLossInputDataSchema(MaBaseSchema):
-    x: NumpyArray = NumpyArray(required=True, allow_none=False)
-    y: NumpyArray = NumpyArray(required=True, allow_none=False)
-    x0: NumpyArray = NumpyArray(required=True, allow_none=False)
-
-    @ma.post_load
-    def make_object(self, data, **kwargs):
-        return CalcLossInputData(**data)
-
-
-@dataclass
-class ObjectiveFunctionCallbackData:
-    calc_loss_endpoint_url: str
-
-
-class ObjectiveFunctionCallbackSchema(MaBaseSchema):
-    calc_loss_endpoint_url = ma.fields.Url(required=True, allow_none=False)
-
-    @ma.post_load
-    def make_object(self, data, **kwargs):
-        return ObjectiveFunctionCallbackData(**data)
-
-
-@dataclass
-class MinimizerCallbackData:
-    minimize_endpoint_url: str
-
-
-class MinimizerCallbackSchema(MaBaseSchema):
-    minimize_endpoint_url = ma.fields.Url(required=True, allow_none=False)
-
-    @ma.post_load
-    def make_object(self, data, **kwargs):
-        return MinimizerCallbackData(**data)
-
-
-@dataclass
-class MinimizerInputData:
-    x: np.ndarray
-    y: np.ndarray
-    calc_loss_endpoint_url: str
-    callback_url: Optional[str] = None
-
-
-class MinimizerInputSchema(MaBaseSchema):
-    x = NumpyArray(required=True, allow_none=False)
-    y = NumpyArray(required=True, allow_none=False)
-    calc_loss_endpoint_url = ma.fields.Url(required=True, allow_none=False)
-    callback_url = ma.fields.Url(required=False, allow_none=False)
-
-    @ma.post_load
-    def make_object(self, data, **kwargs):
-        return MinimizerInputData(**data)
-
-
-@dataclass
-class MinimizerResult:
-    weights: np.ndarray
-
-
-class MinimizerResultSchema(MaBaseSchema):
-    weights = NumpyArray(required=True, allow_none=False)
-
-    @ma.post_load
-    def make_object(self, data, **kwargs):
-        return MinimizerResult(**data)
